factored out login code
